Enable switching camera in SetupActivity
