+++ AzureCalling/app/src/main/java/com/azure/samples/communication/calling/external/calling/CallingContext.java
@@ -235,9 +235,13 @@ public class CallingContext {
                 desiredCamera = getFrontCamera();
             }
             localVideoStream.switchSource(desiredCamera).thenRun(() -> {
-                call.startVideo(appContext, localVideoStream).thenRun(() -> {
+                if (call != null) {
+                    call.startVideo(appContext, localVideoStream).thenRun(() -> {
+                        localVideoStreamCompletableFuture.complete(localVideoStream);
+                    });
+                } else {
                     localVideoStreamCompletableFuture.complete(localVideoStream);
-                });
+                }
             });
         });
     }
