+++ AzureCalling/app/src/main/java/com/azure/samples/communication/calling/view/ParticipantView.java
@@ -8,18 +8,15 @@ import android.graphics.drawable.Drawable;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
-import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintLayout;
-import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.core.content.ContextCompat;
 
 import com.azure.android.communication.calling.CallingCommunicationException;
 import com.azure.android.communication.calling.CreateViewOptions;
-import com.azure.android.communication.calling.LocalVideoStream;
 import com.azure.android.communication.calling.MediaStreamType;
 import com.azure.android.communication.calling.RemoteVideoStream;
 import com.azure.android.communication.calling.VideoStreamRenderer;
@@ -29,17 +26,15 @@ import com.azure.samples.communication.calling.R;
 
 public class ParticipantView extends RelativeLayout {
     // participant view properties
+    protected String videoStreamId;
     private VideoStreamRenderer renderer;
     private VideoStreamRendererView rendererView;
-    private String videoStreamId;
 
     // layout properties
     private final TextView title;
     private final ImageView defaultAvatar;
     private final ConstraintLayout videoContainer;
     private final FrameLayout activeSpeakerFrame;
-    private final ImageButton switchCameraButton;
-    private Runnable switchCameraOnClickAction;
 
     public ParticipantView(@NonNull final Context context) {
         super(context);
@@ -48,13 +43,6 @@ public class ParticipantView extends RelativeLayout {
         this.defaultAvatar = findViewById(R.id.default_avatar);
         this.videoContainer = findViewById(R.id.video_container);
         this.activeSpeakerFrame = findViewById(R.id.active_speaker_frame);
-        this.switchCameraButton = findViewById(R.id.participant_switch_camera_button);
-        switchCameraButton.setOnClickListener(l -> {
-            if (switchCameraOnClickAction != null) {
-                switchCameraButton.setEnabled(false);
-                switchCameraOnClickAction.run();
-            }
-        });
     }
 
     public void setVideoStream(final RemoteVideoStream remoteVideoStream) {
@@ -84,26 +72,6 @@ public class ParticipantView extends RelativeLayout {
         }
     }
 
-    public void setVideoStream(final LocalVideoStream localVideoStream) {
-        if (localVideoStream == null) {
-            cleanUpVideoRendering();
-            return;
-        }
-
-        final String newVideoStreamId = "LocalVideoStream:" + localVideoStream.getSource().getId();
-        if (newVideoStreamId.equals(videoStreamId)) {
-            return;
-        }
-
-        try {
-            final VideoStreamRenderer videoRenderer = new VideoStreamRenderer(localVideoStream, getContext());
-            setVideoRenderer(videoRenderer);
-            this.videoStreamId = newVideoStreamId;
-        } catch (final CallingCommunicationException e) {
-            e.printStackTrace();
-        }
-    }
-
     public void setDisplayName(final String displayName) {
         this.title.setText(displayName);
     }
@@ -127,34 +95,7 @@ public class ParticipantView extends RelativeLayout {
         defaultAvatar.setVisibility(isDisplayVideo ? INVISIBLE : VISIBLE);
     }
 
-    public void setSwitchCameraButtonDisplayed(final boolean shouldShowButton) {
-        switchCameraButton.setVisibility(shouldShowButton ? VISIBLE : GONE);
-    }
-
-    public void setSwitchCameraButtonEnabled(final boolean shouldEnable) {
-        switchCameraButton.setEnabled(shouldEnable);
-    }
-
-    public void centerSwitchCameraButton(final boolean shouldCenter) {
-        final ConstraintSet set = new ConstraintSet();
-        final ConstraintLayout layout;
-
-        layout = (ConstraintLayout) findViewById(R.id.video_container);
-        set.clone(layout);
-        if (shouldCenter) {
-            set.connect(switchCameraButton.getId(), ConstraintSet.BOTTOM, layout.getId(), ConstraintSet.BOTTOM,
-                    4);
-        } else {
-            set.clear(switchCameraButton.getId(), ConstraintSet.BOTTOM);
-        }
-        set.applyTo(layout);
-    }
-
-    public void setImageButtonOnClickAction(final Runnable onClickAction) {
-        switchCameraOnClickAction = onClickAction;
-    }
-
-    private void setVideoRenderer(final VideoStreamRenderer videoRenderer,
+    protected void setVideoRenderer(final VideoStreamRenderer videoRenderer,
                                   final boolean isScreenSharing) {
         this.renderer = videoRenderer;
         this.rendererView = videoRenderer.createView(
@@ -162,7 +103,7 @@ public class ParticipantView extends RelativeLayout {
         attachRendererView(rendererView);
     }
 
-    private void setVideoRenderer(final VideoStreamRenderer videoRenderer) {
+    protected void setVideoRenderer(final VideoStreamRenderer videoRenderer) {
         this.renderer = videoRenderer;
         this.rendererView = videoRenderer.createView(new CreateViewOptions(ScalingMode.CROP));
         attachRendererView(rendererView);
@@ -175,7 +116,6 @@ public class ParticipantView extends RelativeLayout {
             detachFromParentView(rendererView);
             rendererView.setId(View.generateViewId());
             this.videoContainer.addView(rendererView, 0);
-            switchCameraButton.setEnabled(true);
         } else {
             this.defaultAvatar.setVisibility(View.VISIBLE);
         }
