+++ AzureCalling/app/src/main/res/layout/activity_setup.xml
@@ -26,6 +26,24 @@
         android:clickable="true"
         app:layout_constraintBottom_toTopOf="@id/setup_name_holder">
 
+        <ImageButton
+            android:id="@+id/setup_switch_camera_button"
+            android:layout_width="24dp"
+            android:layout_height="24dp"
+            android:layout_marginTop="4dp"
+            android:layout_marginEnd="4dp"
+            android:background="@drawable/button_filled"
+            android:backgroundTint="#99000000"
+            android:contentDescription="@string/switch_camera_button"
+            android:paddingStart="2dp"
+            android:paddingTop="2dp"
+            android:paddingEnd="2dp"
+            android:paddingBottom="2dp"
+            android:visibility="gone"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@drawable/ic_fluent_camera_switch_24_regular"/>
+
         <ProgressBar
             android:id="@+id/setup_progress_bar"
             style="?android:attr/progressBarStyle"
